+++ NEWS.md
@@ -1,6 +1,7 @@
 ## RAnEnExtra 0.2.10
 
 - Added new functio `RAnEnExtra::subsetParameters`
+- Added functionality to parse verification with bootstrap
 
 ## RAnEnExtra 0.2.9